(WIP) creating results mockup
