+++ src/styles/results.scss
@@ -1,0 +1,36 @@
+.results-container {
+  padding: 25px 50px;
+
+  .result-header {
+    padding-bottom: 25px;
+  }
+
+  .result-card {
+    margin-left: auto;
+    margin-right: auto;
+    background-color: #fff;
+    margin-top: 15px;
+    padding: 15px 40px;
+    border-radius: 15px;
+    width: 70%;
+
+    .travel-info {
+      text-align: left;
+      position: relative;
+      width: 70%;
+
+      ul {
+        list-style-type: none;
+        padding-left: 0px;
+        width: 70%;
+        margin: 0px;
+      }
+    }
+
+    .price-info {
+      text-align: right;
+      position: relative;
+      width: 30%;
+    }
+  }
+}
